Clean up numerous warnings on clang and gcc

+++ CoreThreadTest.cpp
@@ -144,7 +144,7 @@ TEST_F(CoreThreadTest, VerifyDispatchQueueShutdown) {
   }
   catch (...) {}
 
-  ASSERT_EQ(listener->GetDispatchQueueLength(), 0);
+  ASSERT_EQ(listener->GetDispatchQueueLength(), static_cast<size_t>(0));
 }
 
 TEST_F(CoreThreadTest, VerifyNoLeakOnExecptions) {
